--- project-generated-code-backend/Backend/Events/Dto/EventParams.cs
@@ -1,6 +1,0 @@
-﻿namespace HealthClinicBackend.Backend.Events.Dto
-{
-    public interface EventParams
-    {
-    }
-}
