--- project-generated-code-backend/Backend/Events/Dto/PatientRegisteredEventParams.cs
@@ -1,7 +1,0 @@
-﻿namespace HealthClinicBackend.Backend.Events.Dto
-{
-    public class PatientRegisteredEventParams : EventParams
-    {
-        public int PatientAge { get; set; }
-    }
-}
