+++ project-generated-code-backend/Backend/Events/EventLogging/Event.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Text;
 
-namespace HealthClinicBackend.Backend.Events.Model
+namespace HealthClinicBackend.Backend.Events.EventLogging
 {
     public class Event
     {
