+++ project-generated-code-backend/Backend/Events/EventLogging/EventParams.cs
@@ -1,0 +1,6 @@
+﻿namespace HealthClinicBackend.Backend.Events.EventLogging
+{
+    public interface EventParams
+    {
+    }
+}
