+++ project-generated-code-backend/Backend/Events/EventLogging/IEventRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace HealthClinicBackend.Backend.Events.EventLogging
+{
+    public interface IEventRepository<in T> where T : EventLogging.Event
+    {
+        void LogEvent(T @event);
+    }
+}
