+++ project-generated-code-backend/Backend/Events/EventLogging/ILogEventService.cs
@@ -1,0 +1,7 @@
+﻿namespace HealthClinicBackend.Backend.Events.EventLogging
+{
+    interface ILogEventService<in T> where T : EventParams
+    {
+        public void LogEvent(T eventParams);
+    }
+}
