--- project-generated-code-backend/Backend/Events/Model/PatientRegisteredEvent.cs
@@ -1,7 +1,0 @@
-﻿namespace HealthClinicBackend.Backend.Events.Model
-{
-    public class PatientRegisteredEvent: Event
-    {
-        public int PatientAge { get; set; }
-    }
-}
