+++ project-generated-code-backend/Backend/Events/PatientRegisteredEventLogging/IPatientRegisteredEventRepository.cs
@@ -1,0 +1,8 @@
+﻿using HealthClinicBackend.Backend.Events.EventLogging;
+
+namespace HealthClinicBackend.Backend.Events.PatientRegisteredEventLogging
+{
+    public interface IPatientRegisteredEventRepository : IEventRepository<PatientRegisteredEvent>
+    {
+    }
+}
