+++ project-generated-code-backend/Backend/Events/PatientRegisteredEventLogging/PatientRegisteredEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace HealthClinicBackend.Backend.Events.PatientRegisteredEventLogging
+{
+    public class PatientRegisteredEvent: EventLogging.Event
+    {
+        public int PatientAge { get; set; }
+    }
+}
