+++ project-generated-code-backend/Backend/Events/PatientRegisteredEventLogging/PatientRegisteredEventDatabase.cs
@@ -1,0 +1,17 @@
+﻿using HealthClinicBackend.Backend.Events.EventLogging;
+
+namespace HealthClinicBackend.Backend.Events.PatientRegisteredEventLogging
+{
+    public class PatientRegisteredEventDatabase : EventDatabase<PatientRegisteredEventLogging.PatientRegisteredEvent>, IPatientRegisteredEventRepository
+    {
+        public PatientRegisteredEventDatabase(EventDbContext dbContext) : base(dbContext)
+        {
+        }
+
+        public override void LogEvent(PatientRegisteredEventLogging.PatientRegisteredEvent @event)
+        {
+            DbContext.PatientRegisteredEvents.Add(@event);
+            DbContext.SaveChanges();
+        }
+    }
+}
