+++ project-generated-code-backend/Backend/Events/PatientRegisteredEventLogging/PatientRegisteredEventParams.cs
@@ -1,0 +1,9 @@
+﻿using HealthClinicBackend.Backend.Events.EventLogging;
+
+namespace HealthClinicBackend.Backend.Events.PatientRegisteredEventLogging
+{
+    public class PatientRegisteredEventParams : EventParams
+    {
+        public int PatientAge { get; set; }
+    }
+}
