--- project-generated-code-backend/Backend/Events/Repository/IEventRepository.cs
@@ -1,12 +1,0 @@
-﻿using HealthClinicBackend.Backend.Events.Model;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace HealthClinicBackend.Backend.Events.Repository
-{
-    public interface IEventRepository<in T> where T : Event
-    {
-        void LogEvent(T @event);
-    }
-}
