--- project-generated-code-backend/Backend/Events/Repository/IPatientRegisteredEventRepository.cs
@@ -1,8 +1,0 @@
-﻿using HealthClinicBackend.Backend.Events.Model;
-
-namespace HealthClinicBackend.Backend.Events.Repository
-{
-    public interface IPatientRegisteredEventRepository : IEventRepository<PatientRegisteredEvent>
-    {
-    }
-}
