--- project-generated-code-backend/Backend/Events/Repository/PatientRegisteredEventDatabase.cs
@@ -1,17 +1,0 @@
-﻿using HealthClinicBackend.Backend.Events.Model;
-
-namespace HealthClinicBackend.Backend.Events.Repository
-{
-    public class PatientRegisteredEventDatabase : EventDatabase<PatientRegisteredEvent>, IPatientRegisteredEventRepository
-    {
-        public PatientRegisteredEventDatabase(EventDbContext dbContext) : base(dbContext)
-        {
-        }
-
-        public override void LogEvent(PatientRegisteredEvent @event)
-        {
-            DbContext.PatientRegisteredEvents.Add(@event);
-            DbContext.SaveChanges();
-        }
-    }
-}
