--- project-generated-code-backend/Backend/Events/Service/ILogEventService.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using HealthClinicBackend.Backend.Events.Dto;
-
-namespace HealthClinicBackend.Backend.Events.Service
-{
-    interface ILogEventService<in T> where T : EventParams
-    {
-        public void LogEvent(T eventParams);
-    }
-}
